[Refactor] Extract DataLoaderMock into a new file
